requirements.txt
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
numpy>=1.24.0
ollama>=0.1.0 